app/requirements.txt
absl-py==1.4.0
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.23.0
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.15.2
google-cloud-resource-manager==1.9.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.48.2
h5py==3.8.0
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.0.4
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
stack-data==0.6.2
streamlit==1.20.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
zipp==3.15.0
